+++ lib/Text/ASCIITable/EasyTable.pm
@@ -13,16 +13,39 @@ use parent qw(Exporter);
 
 our @EXPORT = qw(easy_table);  ## no critic (ProhibitAutomaticExportation)
 
-our $VERSION = '1.003';
+our $VERSION = '1.004';
 
 ########################################################################
-sub is_array { push @_, 'ARRAY'; goto &_is_type; }
-sub is_hash  { push @_, 'HASH';  goto &_is_type; }
+{
+  ## no critic (RequireArgUnpacking)
+
+  sub is_array { push @_, 'ARRAY'; goto &_is_type; }
+  sub is_hash  { push @_, 'HASH';  goto &_is_type; }
+  sub _is_type { return ref $_[0] && reftype( $_[0] ) eq $_[1]; }
+}
 ########################################################################
 
 ########################################################################
-sub _is_type { return ref $_[0] && reftype( $_[0] ) eq $_[1]; }
+sub uncamel {
 ########################################################################
+  my ($str) = @_;
+
+  while ( $str =~ s/^(.)(.*?)([[:upper:]])/\l$1$2_\l$3/xsmg ) { }
+
+  return $str;
+}
+
+########################################################################
+sub wordify {
+########################################################################
+  my ($str) = @_;
+
+  $str = uncamel($str);
+
+  $str =~ s/_(.)/ \u$1/xsmg;
+
+  return ucfirst $str;
+}
 
 ########################################################################
 sub easy_table {
@@ -99,7 +122,13 @@ sub _render_table {
 
   my $t = Text::ASCIITable->new($table_options);
 
-  $t->setCols( @{ $options{columns} } );
+  my @columns = @{ $options{columns} };
+
+  if ( $options{fix_headings} ) {
+    @columns = map { wordify $_ } @columns;
+  }
+
+  $t->setCols(@columns);
 
   for ( @{ $options{data} } ) {
     $t->addRow( @{$_} );
@@ -186,15 +215,15 @@ Text::ASCIITable::EasyTable - create ASCII tables from an array of hashes
  
  print easy_table(
    data          => $data,
-   index         => \@index,
-   table_options => { headerText => 'My Easy Table' },
+   rows          => $rows,
+   table_options => { headingText => 'My Easy Table' },
  );
 
  # easier 
  print easy_table(
    data          => $data,
    columns       => [ sort keys %{ $data->[0] } ],
-   table_options => { headerText => 'My Easy Table' },
+   table_options => { headingText => 'My Easy Table' },
  );
  
  # easiest 
@@ -222,8 +251,9 @@ also allow you to set the order of the data to be displayed in the table.
 
 =item * Sort rows by individual columns in the hashes
 
-=item * Output JSON instead of a tableInstead of rendering a table, C<easy_table> can apply the same type of
-transformations to arrays of hashes and subsequently output JSON.
+=item * Output JSON instead of a tableInstead of rendering a table,
+C<easy_table> can apply the same type of transformations to arrays of
+hashes and subsequently output JSON.
 
 =back
 
@@ -357,6 +387,20 @@ to transform the column names and column values in a table.
 =item max_rows
 
 Maximum number of rows to render.
+
+=item fix_headings
+
+Many data sets will contain hash keys composed of lower case letters
+in what is termed I<snake case> (words separated by '_') or I<camel
+case> (first letter of words in upper case). Set this to true to turn
+snake and camel case into space separated 'ucfirst'ed words.
+
+Example:
+
+ creation_date => Creation Date
+ IsTruncated   => Is Truncated
+
+default: false
 
 =item sort_key